[ALPS02353522] DISP: coverity fix
[Detail]
coverity fix in primary_display.c

MTK-Commit-Id: 16eb23ebcaa40e11ef8c5be4e264873cac0284d8

CR-Id: ALPS02353522
Feature: Display Driver (DSI Interface)

+++ drivers/misc/mediatek/video/mt6739/videox/primary_display.c
@@ -6221,7 +6221,8 @@ static int smart_ovl_try_switch_mode_nolock(void)
 	data_config = dpmgr_path_get_last_config(disp_handle);
 
 	/* calc wdma/rdma data size */
-	rdma_sz = data_config->dst_h * data_config->dst_w * 3;
+	rdma_sz = (unsigned long long)data_config->dst_h *
+				      data_config->dst_w * 3;
 
 	/* calc ovl data size */
 	ovl_sz = 0;
